ai-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0.0
semantic-kernel>=1.0.0
azure-cognitiveservices-speech
azure-ai-textanalytics
langdetect
pydantic
httpx